subscribe to message events
