+++ packages/slack/src/mention.ts
@@ -1,4 +1,7 @@
+import { formatAPIError, getAPIErrorMessage } from '@globalping/bot-utils';
+import { WebClient } from '@slack/web-api';
 
+import { logger, postAPI } from './utils';
 
 export function parseCommandfromMention(text: string, botUserId: string): string {
     const trimmedText = text.trim();
@@ -12,4 +15,23 @@ export function parseCommandfromMention(text: string, botUserId: string): string
     }
 
     return '';
+}
+
+export async function handleMention(fullText: string, teamId: string | undefined, channelId: string, userId: string, eventTs: string, threadTs: string | undefined, botUserId: string, client: WebClient) {
+    const logData = { fullText, teamId, channelId, userId, eventTs, threadTs };
+    logger.info(logData, '@globalping request');
+
+    const commandText = parseCommandfromMention(fullText, botUserId);
+
+    try {
+        // the mention is always received in a channel where the bot is a member
+        const channelPayload = { channel_id: channelId, user_id: userId, thread_ts: threadTs };
+        logger.info({ commandText, ...logData }, '@globalping processing starting');
+        await postAPI(client, channelPayload, commandText);
+        logger.info(logData, '@globalping response - OK');
+    } catch (error) {
+        const errorMsg = getAPIErrorMessage(error);
+        logger.error({ errorMsg, ...logData }, '@globalping failed');
+        await client.chat.postMessage({ channel: channelId, thread_ts: threadTs, text: `Failed to process command \`${commandText}\`.\n${formatAPIError(errorMsg)}` });
+    }
 }
